fixed logic

+++ container/prepare.sh
@@ -76,11 +76,32 @@ mkdir -p "${BASE_DIR}/${WORK_DIR}"
 
 
 ###############################
+# Container scripts
+echo ">>>> running container scripts"
+cd "${BASE_DIR}"
+chmod +x *.sh
+if [ ! -f pull-zowe-install-artifacts.sh ]; then
+  echo "Error: pull-zowe-install-artifacts script is missing."
+  exit 3
+fi
+./pull-zowe-install-artifacts.sh
+
+if [ ! -f download-zlux.sh ]; then
+  echo "Error: download-zlux script is missing."
+  exit 4
+fi
+./download-zlux.sh
+
+
+
+###############################
 echo ">>>>> prepare basic files"
 cd "${REPO_ROOT_DIR}"
 cp README.md "${BASE_DIR}/${WORK_DIR}"
 cp LICENSE "${BASE_DIR}/${WORK_DIR}"
 cp package.json "${BASE_DIR}/${WORK_DIR}"
+
+cp -rf files "${BASE_DIR}/${WORK_DIR}"
 
 
 ###############################
@@ -107,23 +128,6 @@ cat manifest.yaml | \
 # copy to target context
 echo ">>>>> copy to target build context"
 cp -r "${BASE_DIR}/${WORK_DIR}" "${BASE_DIR}/${linux_distro}/${cpu_arch}/component"
-
-###############################
-# Container scripts
-echo ">>>> running container scripts"
-cd "${BASE_DIR}"
-chmod +x *.sh
-if [ ! -f pull-zowe-install-artifacts.sh ]; then
-  echo "Error: pull-zowe-install-artifacts script is missing."
-  exit 3
-fi
-./pull-zowe-install-artifacts.sh
-
-if [ ! -f download-zlux.sh ]; then
-  echo "Error: download-zlux script is missing."
-  exit 4
-fi
-./download-zlux.sh
 
 
 ###############################
